working on oce-ocedata split
I see no solution to this problem.  My present thought is to use
ocedata only for large datasets, and to merge the small ones back into
oce itself.  If I put just the following here in ocedata, oce will be
smaller than it was before, but still workable.  (PS. I can make adp
smaller, and should do that because it belongs in oce, not ocedata, it
being a core dataset.)

     3.0M  data/coastlineWorldFine.rda
    490K  data/coastlineWorld.rda
    377K  data/adp.rda
    299K  data/topoWorld.rda
    194K  data/levitus.rda
    177K  data/secchi.rda