undo URL change in DESCRIPTION
What I read in Rjournal just doesn't seem to be true; for *some* (but
not all) instances, R-build-check gives a warning if the URL is enclosed
in <>, as is recommended.  I checked with the MASS source DESCRIPTION to
try to learn more, but it does not have a bugreports field, and that's
where I got the problem.