fix R-CMD-check workflow(?); eliminate pkgdown workflow
Why eliminate the pkgdown workflow?  Because I often run into problems
with it, and, besides, I prefer building locally and looking at the
webpage, in case I see the need for changes.

Github workflows are a pain.  I get that they are now the expected thing,
but when have they helped me to find bugs?  Answer: never, as far as I
can recall.  Local tests, and a manually-started external windows build,
seem to work just fine.